Add files via upload
Simple practice of inheritance and general OOP. Need to up my programming with objects. The most interesting way for me is to think of it like a video game where you have friendly characters and enemy characters and other sub-objects. Don't really know if this is best practice or efficient code but you know, it never really is...